cleaners finis

+++ src/GPS.h
@@ -11,10 +11,10 @@ public :
 
     GPS(double latitude, double longitude);
 
-    inline int get_latitude() const {
+    inline double get_latitude() const {
         return latitude;
     }
-    inline int get_longitude() const {
+    inline double get_longitude() const {
         return longitude;
     }
